flamegraph: Refactor JavaScript into two modules
This lays the ground work for us to reuse our flame graph generation
tools across multiple different types of flame graphs.

Every function and class deleted from `flamegraph.js` is added without
any changes (other than maybe an `export`) to `flamegraph_common.js`.
