+++ Makefile
@@ -13,6 +13,7 @@ PIP_INSTALL=PKG_CONFIG_PATH="$(PKG_CONFIG_PATH)" $(PYTHON) -m pip install
 reporters_path := ./src/memray/reporters
 js_files := $(wildcard $(reporters_path)/assets/*.js)
 generated_js_files := \
+    $(reporters_path)/templates/assets/flamegraph_common.js \
     $(reporters_path)/templates/assets/flamegraph.js \
     $(reporters_path)/templates/assets/table.js
 css_files := 'src/**/*.css'
