+++ pyproject.toml
@@ -40,8 +40,7 @@ xfail_strict = true
 
 [tool.check-manifest]
 ignore = [
-  "src/memray/reporters/templates/assets/flamegraph.js",
-  "src/memray/reporters/templates/assets/table.js",
+  "src/memray/reporters/templates/assets/*.js",
 ]
 
 [tool.mypy]
