+++ webpack.config.js
@@ -4,6 +4,7 @@ const TerserPlugin = require("terser-webpack-plugin");
 module.exports = {
   mode: "production",
   entry: {
+    flamegraph_common: "./src/memray/reporters/assets/flamegraph_common.js",
     flamegraph: "./src/memray/reporters/assets/flamegraph.js",
     table: "./src/memray/reporters/assets/table.js",
   },
